requirements.txt
firebase_admin==6.5.0
Flask==3.0.3
Flask_Cors==4.0.1
PyJWT==2.8.0
PyJWT==2.8.0
python_bcrypt==0.3.2
python_gemini_api==2.4.12
Requests==2.32.3
